Implement new preview sync mechanism
Current mechanism keeps track of the precise time a frame has been
shown to the user via a callback right after glfwSwapBuffers.

When a new frame was fetched and is to be displayed, we tell the MainWindow
the exact "frame_sync_time" when the previously shown frame should end.

It is then up to the MainWindow to figure out if it wants to sleep
to ensure proper frame duration.

Becuase the decision to retire a frame is made based on absolute
timestamp difference instead of conditional delta time accumulation
we still have the issue where time when paused is counted to frame
display time.

This effect is now confined to only one frame, hence it's not noticable
enough to be a nuisance for normal video (30-60FPS), although it might lead to
subtle bugs where quickly toggling the playback state of a low FPS video (~8FPS)
can skip frames prematurely.